+++ mr/HH.hpp
@@ -38,8 +38,8 @@ namespace mr
   class HH<OS> : public PoleMassAndCouplings
   {
 
-    long double MMb, MMt, MMH, MMW, MMZ, mu2;
-    long double SW, CW;
+    double MMb, MMt, MMH, MMW, MMZ, mu2;
+    double SW, CW;
   
     std::unique_ptr<Tsil> protHHHHH;
     std::unique_ptr<Tsil> protHZHZZ;
@@ -70,57 +70,57 @@ namespace mr
     }
 
 
-    HH(long double,long double,long double,long double,long double);
+    HH(double,double,double,double,double);
   
-    HH(OSinput, long double);
+    HH(OSinput, double);
   
     void init();
 
-    long double Q()
+    double Q()
     {
       return sqrt(mu2);
     }
 
-    long double QQ()
+    double QQ()
     {
       return mu2;
     }
   
   
-    long double x10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double x10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
   
-    long double x11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double x11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
   
 
-    long double x20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double x20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
 
     // Gaugeless limit
-    long double xgl10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double xgl10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
   
-    long double xgl11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double xgl11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
-    long double xgl20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double xgl20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
 
     
     // 
     // \lambda/sqrt(2)/GF/MMH
     // 
-    long double y10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double y10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
-    long double y11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double y11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
   
-    long double y20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double y20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
     // Gaugeless limit
 
-    long double ygl10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double ygl10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
-    long double ygl11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double ygl11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
-    long double ygl20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double ygl20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
   
   };
 
@@ -129,8 +129,8 @@ namespace mr
   class HH<MS> : public BaseMass
   {
 
-    long double mmb, mmt, mmH, mmW, mmZ, mu2;
-    long double s, c;
+    double mmb, mmt, mmH, mmW, mmZ, mu2;
+    double s, c;
   
     std::unique_ptr<Tsil> protHHHHH;
     std::unique_ptr<Tsil> protHZHZZ;
@@ -161,22 +161,22 @@ namespace mr
     }
 
 
-    HH(long double,long double,long double,long double,long double);
+    HH(double,double,double,double,double);
   
-    HH(MSinput, long double);
+    HH(MSinput, double);
   
     void init();
 
-    //  long double x01(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    //  double x01(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
   
-    long double x10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double x10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
   
-    long double x11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double x11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
   
 
-    long double x20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double x20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
 
   };
