+++ mr/WW.hpp
@@ -39,8 +39,8 @@ namespace mr
   class WW<OS> : public PoleMassAndCouplings
   {
 
-    long double MMb, MMt, MMH, MMW, MMZ, mu2;
-    long double SW, CW;
+    double MMb, MMt, MMH, MMW, MMZ, mu2;
+    double SW, CW;
 
     std::unique_ptr<Tsil> protWHHWW;
     std::unique_ptr<Tsil> protWHZWW;
@@ -72,32 +72,32 @@ namespace mr
     {
     }
 
-    WW(long double,long double,long double,long double,long double);
+    WW(double,double,double,double,double);
 
-    WW(OSinput, long double);
+    WW(OSinput, double);
 
     void init();
   
 
     // Pole -> MS
   
-    long double x10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double x10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
   
-    long double x11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double x11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
   
 
-    long double x20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double x20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
 
     // Yukawa
-    long double y10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double y10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
   
-    long double y11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double y11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
   
 
-    long double y20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double y20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
   };
 
@@ -106,8 +106,8 @@ namespace mr
   class WW<MS> : public BaseMass
   {
 
-    long double mmb, mmt, mmH, mmW, mmZ, mu2;
-    long double s, c;
+    double mmb, mmt, mmH, mmW, mmZ, mu2;
+    double s, c;
 
     std::unique_ptr<Tsil> protWHHWW;
     std::unique_ptr<Tsil> protWHZWW;
@@ -139,21 +139,21 @@ namespace mr
     {
     }
 
-    WW(long double,long double,long double,long double,long double);
+    WW(double,double,double,double,double);
 
-    WW(MSinput, long double);
+    WW(MSinput, double);
 
     void init();
   
     // Pole -> MS
   
-    long double x10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double x10(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
   
-    long double x11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double x11(size_t nL = 2, size_t nH = 1, size_t boson = 1);
   
 
-    long double x20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
+    double x20(size_t nL = 2, size_t nH = 1, size_t boson = 1);
 
 
   };
